feat: change up EditMenu

+++ MediaManager/MediaManager/GUI/Menus/EditMenu.xaml.cs
@@ -20,42 +20,164 @@ namespace MediaManager.GUI.Menus
     {
         #region Setup
         private int MediumId;
-        private int? SelectedPartId;
+        private bool AnyChangeMade = false;
 
-        public EditMenu(int? mediumId, int? partId)
+        /// <summary>
+        /// Opens the editor or viewer for the given medium.
+        /// </summary>
+        /// <param name="mediumId">Medium to open - creates a new one if null</param>
+        /// <param name="partId">Part within the medium to open - only possible if medium exists</param>
+        /// <param name="edit">Whether to directly start editing - will always be the case for new media</param>
+        public EditMenu(int? mediumId, int? partId, bool edit)
         {
             InitializeComponent();
             RegisterAtLanguageProvider();
-            updateElementChangeOptions(true);
+            MediumId = mediumId ?? -1;
+            updateVisibility(edit);
 
-            MediumId = mediumId ?? Writer.CreateMedium();
-            SelectedPartId = mediumId != null ? partId : null;
-            reloadData();
-            if (!mediumId.HasValue) startEditing(); // directly edit new medium
+            if (IsExistingMedium)
+            {
+                reloadData();
+                if (partId.HasValue)
+                {
+                    viewer.OpenPartTab(partId.Value);
+                    editor.OpenPartTab(partId.Value);
+                }
+            }
+            else
+            {
+                editor.Medium = new Controls.Edit.MediumWithTags
+                {
+                    Id = MediumId,
+                    CatalogueId = CURRENT_CATALOGUE.Id,
+                    Title = "",
+                    Description = "",
+                    Location = "",
+                    Tags = CURRENT_CATALOGUE.Tags.Select(t => new ValuedTag
+                    {
+                        Tag = t,
+                        Value = null
+                    }).ToList(),
+                    Parts = new System.Collections.Generic.List<Controls.Edit.PartWithTags>()
+                };
+                updateVisibility(true); // start editing
+            }
         }
+        private bool IsExistingMedium { get => MediumId >= 0; }
         private void reloadData()
         {
-            var medium = Reader.GetMedium(MediumId);
-            list.SetPartList(medium.Title, medium.Parts.Select(m => new Controls.List.PartListElement
+            if (IsExistingMedium)
             {
-                Id = m.Id,
-                Title = m.Title
-            }).ToList());
-            list.SelectItem(SelectedPartId);
-        }
-        private void updateElementChangeOptions(bool available)
-        {
-            Resources["btnAddPartVisible"] = available ? Visibility.Visible : Visibility.Collapsed;
-            list.IsEnabled = available;
+                var medium = Reader.GetMedium(MediumId);
+                var mediaData = new Controls.Edit.MediumWithTags
+                {
+                    Id = medium.Id,
+                    CatalogueId = medium.CatalogueId,
+                    Title = medium.Title,
+                    Description = medium.Description,
+                    Location = medium.Location,
+                    Tags = Reader.GetTagsForMedium(MediumId),
+                    Parts = medium.Parts.Select(p => new Controls.Edit.PartWithTags
+                    {
+                        Id = p.Id,
+                        Title = p.Title,
+                        Description = p.Description,
+                        Favourite = p.Favourite,
+                        Length = p.Length,
+                        Publication_Year = p.Publication_Year,
+                        Image = p.Image,
+                        Tags = Reader.GetTagsForPart(p.Id)
+                    }).ToList()
+                };
+                viewer.Medium = mediaData;
+                editor.Medium = mediaData;
+            }
         }
 
         public void RegisterAtLanguageProvider() => LanguageProvider.RegisterUnique(this);
         public void LoadTexts(string language)
         {
-            Resources["btnAddPart"] = LanguageProvider.getString("Menus.Edit.ToolTip.AddPart");
             Resources["btnDeleteMedium"] = LanguageProvider.getString("Menus.Edit.ToolTip.DeleteMedium");
         }
         #endregion
+
+        private bool validate() => RunValidation(new System.Collections.Generic.List<Func<string>>
+        {
+            () => editor.Medium.Title.Trim().Length == 0 ? LanguageProvider.getString("Menus.Edit.Validation.MediaTitle") : null,
+            () => editor.Medium.Parts.Any(p => p.Title.Trim().Length == 0) ? LanguageProvider.getString("Menus.Edit.Validation.PartTitle") : null,
+            () => editor.Medium.Tags.Where(t => t.Value.HasValue).Any(t => editor.Medium.Parts.Any(p => p.Tags.Find(pt => pt.Tag.Id == t.Tag.Id).Value != t.Value)) ? LanguageProvider.getString("Menus.Edit.Validation.PartTags") : null,
+        });
+        private bool save()
+        {
+            if (!validate()) return false;
+            try
+            {
+                var data = editor.Medium;
+                if (IsExistingMedium)
+                {
+                    Writer.SaveMedium(new Medium
+                    {
+                        Id = MediumId,
+                        CatalogueId = data.CatalogueId,
+                        Title = data.Title,
+                        Description = data.Description,
+                        Location = data.Location
+                    }, data.Tags);
+                    // remove unused parts
+                    var oldParts = Reader.GetMedium(MediumId).Parts;
+                    var newParts = data.Parts.Select(p => p.Id).Where(p => p >= 0).ToList();
+                    oldParts.Where(p => !newParts.Contains(p.Id)).ToList().ForEach(p => Writer.DeletePart(p.Id));
+                    // save updated parts
+                    data.Parts.Where(p => p.Id >= 0).ToList().ForEach(p => Writer.SavePart(new Part
+                    {
+                        Id = p.Id,
+                        MediumId = MediumId,
+                        Title = p.Title,
+                        Description = p.Description,
+                        Favourite = p.Favourite,
+                        Length = p.Length,
+                        Publication_Year = p.Publication_Year,
+                        Image = p.Image,
+                    }, p.Tags));
+                    // add new parts
+                    data.Parts.Where(p => p.Id < 0).ToList().ForEach(p => Writer.CreatePart(new Part
+                    {
+                        MediumId = MediumId,
+                        Title = p.Title,
+                        Description = p.Description,
+                        Favourite = p.Favourite,
+                        Length = p.Length,
+                        Publication_Year = p.Publication_Year,
+                        Image = p.Image,
+                    }, p.Tags));
+                }
+                else
+                {
+                    var mediumId = Writer.CreateMedium(new Medium
+                    {
+                        CatalogueId = data.CatalogueId,
+                        Title = data.Title,
+                        Description = data.Description,
+                        Location = data.Location
+                    }, data.Tags);
+                    data.Parts.ForEach(p => Writer.CreatePart(new Part
+                    {
+                        MediumId = mediumId,
+                        Title = p.Title,
+                        Description = p.Description,
+                        Favourite = p.Favourite,
+                        Length = p.Length,
+                        Publication_Year = p.Publication_Year,
+                        Image = p.Image,
+                    }, p.Tags));
+                }
+                return true;
+            }
+            catch (Exception)
+            {
+                return false;
+            }
+        }
 
         #region Navbar
         private void NavigationBar_BackClicked(object sender, EventArgs e)
@@ -66,27 +188,27 @@ namespace MediaManager.GUI.Menus
                 return;
             }
 
-            // there are possibly unsaved changes => ask user
-            var confirmation = ShowUnsavedChangesDialog();
-            // no value => discard
-            // false => close
-            // true => try saving => close, if successful
-            if (confirmation.HasValue && (!confirmation.Value || (confirmation.Value && editor.saveChanges())))
+            var confirmation = !AnyChangeMade;
+            if (!confirmation)
             {
-                Close();
-                Writer.CleanupEmptyMediaAndParts();
+                var result = ShowUnsavedChangesDialog();
+                confirmation = result.HasValue;
+                // if should save => try saving and cancel closing, if saving fails
+                if (result.HasValue && result.Value && !save()) confirmation = false;
             }
+            if (confirmation) Close();
         }
         private void NavigationBar_HelpClicked(object sender, EventArgs e)
         {
             throw new NotImplementedException();
         }
-        private void btnAddPartClick(object sender, RoutedEventArgs e)
+
+        private void updateVisibility(bool editMode)
         {
-            var newPartId = Writer.CreatePart(MediumId);
-            reloadData();
-            list.SelectItem(newPartId);
-            startEditing();
+            viewer.Visibility = editMode ? Visibility.Collapsed : Visibility.Visible;
+            editor.Visibility = editMode ? Visibility.Visible : Visibility.Collapsed;
+            Resources["viewerButtonsVisibility"] = editMode ? Visibility.Collapsed : Visibility.Visible;
+            Resources["editorButtonsVisibility"] = editMode ? Visibility.Visible : Visibility.Collapsed;
         }
         private void btnDeleteMediumClick(object sender, RoutedEventArgs e)
         {
@@ -98,60 +220,30 @@ namespace MediaManager.GUI.Menus
             }
             if (performDeletion)
             {
-                Writer.DeleteMedium(MediumId);
+                if (IsExistingMedium) Writer.DeleteMedium(MediumId);
                 Close();
             }
         }
-        #endregion
-
-        #region Change View
-        private void list_SelectionChanged(int? id)
+        private void btnEditMediumClick(object sender, RoutedEventArgs e)
         {
-            SelectedPartId = id;
-            viewer.Visibility = Visibility.Visible;
-            editor.Visibility = Visibility.Collapsed;
-            viewer.LoadElement(
-                SelectedPartId.HasValue ? Controls.Edit.ElementMode.Part : Controls.Edit.ElementMode.Medium,
-                SelectedPartId.HasValue ? SelectedPartId.Value : MediumId
-                );
+            editor.Medium = viewer.Medium;
+            updateVisibility(true);
         }
-        private void startEditing()
+        private void btnUndoChangesClick(object sender, RoutedEventArgs e)
         {
-            viewer.Visibility = Visibility.Collapsed;
-            editor.Visibility = Visibility.Visible;
-            editor.LoadElement(
-                SelectedPartId.HasValue ? Controls.Edit.ElementMode.Part : Controls.Edit.ElementMode.Medium,
-                SelectedPartId.HasValue ? SelectedPartId.Value : MediumId
-                );
-            updateElementChangeOptions(false);
+            if (!IsExistingMedium) Close();
+            AnyChangeMade = false;
+            updateVisibility(false);
         }
-
-        private void viewer_EditClicked(Controls.Edit.ElementMode mode, int id)
+        private void btnSaveMediumClick(object sender, RoutedEventArgs e)
         {
-            viewer.Visibility = Visibility.Collapsed;
-            editor.Visibility = Visibility.Visible;
-            editor.LoadElement(mode, id);
-            updateElementChangeOptions(false);
-        }
-        private void viewer_DeleteClicked(Controls.Edit.ElementMode mode, int id)
-        {
-            if (mode == Controls.Edit.ElementMode.Medium) Close();
-            else
-            {
-                list.SelectItem(null);
-                reloadData();
-            }
-            updateElementChangeOptions(true);
-        }
-        private void editor_QuitEditing(Controls.Edit.ElementMode mode, int id)
-        {
-            viewer.Visibility = Visibility.Visible;
-            editor.Visibility = Visibility.Collapsed;
-            viewer.LoadElement(mode, id);
-            var medium = Reader.GetMedium(MediumId);
+            save();
             reloadData();
-            updateElementChangeOptions(true);
+            AnyChangeMade = false;
+            updateVisibility(false);
         }
         #endregion
+
+        private void editor_MediumEdited(Controls.Edit.MediumWithTags medium) => AnyChangeMade = true;
     }
 }
